unmkbootimg: fix alignment of second bootloader
The kernel and ramdisk are aligned separately so adding their sizes
together and aligning the result will be incorrect if each is less than
half a page past a boundary.  While doing this introduce a helper
function and use bitwise operations to more clearly match the
calculation in mkbootimg.c.